torii: Moved the `warnings` module into `diagnostics.hooks`
