+++ tests/diagnostics/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: BSD-2-Clause
