+++ torii/diagnostics/hooks.py
